Moving the psf models to ctapipe.instrument.optics
