+++ docs/changes/2643.feature.rst
@@ -1,4 +1,4 @@
-Adds ``ctapipe.image.psf_model`` with the parent class for psf models called ``PSFModel`` and a psf model based on pure coma abbaration called ``ComaModel``
+Adds psf models to ``ctapipe.instrument.optics`` with the parent class ``PSFModel`` and a psf model based on pure coma abbaration called ``ComaModel``
 - All psf models have a function ``PSFModel.update_location`` that updates the location in the focal plane where the PSF is evaluated from
 - The function ``PSFModel.update_model_parameters`` allows to update the parameters of the PSF model
 - The function ``PSFModel.pdf`` gives the value of the PSF in a given location
